requirements.txt
feedparser
openai
litellm
crawl4ai
playwright
python-dotenv
googlenewsdecoder
